Add preliminary support for BoringSSL as an alternative to OpenSSL
When using BoringSSL:
- Newer ciphers are only available using the dedicated AEAD API.
In order to use ChaChaPoly with distinct init/update/finalize()
functions, we need to reimplement it.
- Blowfish has been deprecated and was removed from libcrypto;
linking the decrepit library is required.
BoringSSL doesn't come with a pkgconfig file, so defining
BORINGSSL_ROOT_DIR is required.
- ChaCha cannot be used as a stream cipher without its dedicated API.

BoringSSL/Linux is also checked on CI.

There are no BoringSSL packages not point releases, so we have to
recompile it from source code.